Refactor: Final api service refactor

+++ CarRentalApi.Domain/Entity/Car.cs
@@ -1,14 +1,14 @@
 ﻿using System;
-using CarRentalApi.Domain.Do;
+using CarRentalApi.Domain.Dto;
 
 namespace CarRentalApi.Domain.Entity
 {
     public class Car
     {
-        public int ID { get; set; }
-        public int CarModelID { get; set; }
+        public long ID { get; set; }
+        public long CarModelID { get; set; }
         public CarModel CarModel { get; set; }
-        public int CategoryID { get; set; }
+        public long CategoryID { get; set; }
         public Category Category { get; set; }
         public int ProductionYear { get; set; }
         public int Capacity { get; set; }
